Build a complete network and helpful artifacts

Build a complete network and helpful artifacts

Usage

build_network(key_actor, keywords, token, refresh_tok, threshold, ...)

Arguments

key_actor

account identifier

keywords

character vector with keywords that we check for in actors' description

token

character, token for Blue Sky Social API

refresh_tok

character, refresh token for Blue Sky Social API

threshold

the threshold for including actors in the expansion: How many followers must a prospect have, to be considered? If smaller than 1, threshold is used as a proportion: How big a proportion must a prospect have, to be considered?

...

Parameters passed on to expand_net()

Value

list of objects generated from expand_net(), trim_net(), get_profiles(), create_widget(), and word_freqs()